Update mainClassProj
Second update, fixed a small bug in my move(). also changed the value of "val" on line 144 from an empty string to a null string, did this because if we use remove with an invalid index and print it out in driver, it'll show "null" as opposed to just nothing (if we used "").